+++ src/widgets/world/RenderSettingsWidget.cpp
@@ -100,6 +100,10 @@ int RenderSettingsWidget::shadowMapSize() const {
   return p->ui.rasterShadowMapSize->value();
 }
 
+int RenderSettingsWidget::shadowCascadeCount() const {
+  return p->ui.rasterShadowCascadeCount->value();
+}
+
 double RenderSettingsWidget::shadowBias() const {
   return p->ui.rasterShadowBias->value();
 }
@@ -150,6 +154,8 @@ void RenderSettingsWidget::updateEngineControls() {
   p->ui.rasterShadowMaps->setVisible(isRasterizer);
   p->ui.label_rasterShadowMapSize->setVisible(showShadowDetails);
   p->ui.rasterShadowMapSize->setVisible(showShadowDetails);
+  p->ui.label_rasterShadowCascadeCount->setVisible(showShadowDetails);
+  p->ui.rasterShadowCascadeCount->setVisible(showShadowDetails);
   p->ui.label_rasterShadowBias->setVisible(showShadowDetails);
   p->ui.rasterShadowBias->setVisible(showShadowDetails);
   p->ui.label_rasterShadowFilterRadius->setVisible(showShadowDetails);
@@ -176,6 +182,7 @@ void RenderSettingsWidget::setBusy(bool busy) {
   p->ui.rasterPostProcessAA->setEnabled(!busy);
   p->ui.rasterShadowMaps->setEnabled(!busy);
   p->ui.rasterShadowMapSize->setEnabled(!busy);
+  p->ui.rasterShadowCascadeCount->setEnabled(!busy);
   p->ui.rasterShadowBias->setEnabled(!busy);
   p->ui.rasterShadowFilterRadius->setEnabled(!busy);
   p->ui.rasterShadowFilterMode->setEnabled(!busy);